requirements.txt
anytree==2.8.0
bottle==0.12.18
pony==0.7.13
python-dateutil==2.8.1
PyYAML==5.3.1
rfeed==1.1.1
six==1.15.0
tqdm==4.51.0
